rewrite faq entry on mandatory parameters
The original uses "mandatory" and then "required." I swapped where
those two are used, putting "required" in the first sentence and
"mandatory in the other." Small change, but I believe it's better. Here's why:

First, we restate the error without using the all-important word,
"mandatory." If the user needs to read the FAQ for this error, it's
likely they're struggling with this word in particular.

Second, we refer the user to the API docs as in the original, but now
we use "mandatory" to remind the user to keep that word in mind as
they read the docs.