rewrite section on postman safety practices
Correct typos, punctuation, and grammar. Standardize spelling. Add
bold text to the warning about withdrawals.

I suggest that this section also link to an explanation of what the
potential dangers of using Postman might be. Those aren't addressed,
but a user who's new to Postman might not be reassured by the "safety
first" language in this section.