Correcting grammar, typos, and other readability issues in the English README  (#44)
* correct and rewrite introduction

This fixes typos in comma use, more closely follows the Chinese
README, and specifies what kind of file a Postman environment is when
that term is first used.

* fix typos in section on using Collections

Standardize spelling and correct typos. In particular, collection ->
Collection and APP -> app.

* correct environment file tutorial - initial steps

Standardize spelling of JSON and Postman. Add punctuation.

* rewrite environment import tutorial - manage environments

Correct grammar ("button" should follow its name), spacing, and
punctuation. Following the Chinese README, clarify that Postman for Mac is used in this example.

* rewrite environment import tutorial - importing

Explicitly mention the Add button. Standardize naming of buttons to
use "Capital Case." Correct spacing and punctuation.

* rewrite environment import tutorial - setting values

The original README said "don't set Initial Value" on one line, and
"don't leave Initial Value empty" on the next. These instructions were
contradictory, so something had to be fixed.

I found that the Chinese README actually says that the user should set Current Value...

> 记得输入到`Current Value`里面，不然postman会上传到你的postman账号里面。

...so I followed the Chinese, and rewrote this step's reminder in
terms of Postman's Current Value column (and bolded the warning part.)

Then, I moved the other notes (including the one about auto-filling
Initial Value) to the end of the step.

* rewrite environment import tutorial - picking environment

Clarify where the environment file dropdown menu is found on Mac.

* add alt tags to HTML images for accessibility

* add prose sentence explaining the purpose of the hyperlink

* rewrite faq entry

Correct typos and grammar.

* rewrite faq entry about invalid API keys

Correct grammar.

* rewrite faq entry about signatures

Correct grammar.

* rewrite faq entry on mandatory parameters

The original uses "mandatory" and then "required." I swapped where
those two are used, putting "required" in the first sentence and
"mandatory in the other." Small change, but I believe it's better. Here's why:

First, we restate the error without using the all-important word,
"mandatory." If the user needs to read the FAQ for this error, it's
likely they're struggling with this word in particular.

Second, we refer the user to the API docs as in the original, but now
we use "mandatory" to remind the user to keep that word in mind as
they read the docs.

* fix faq about request urls

Correct grammar and punctuation.

* rewrite section on postman safety practices

Correct typos, punctuation, and grammar. Standardize spelling. Add
bold text to the warning about withdrawals.

I suggest that this section also link to an explanation of what the
potential dangers of using Postman might be. Those aren't addressed,
but a user who's new to Postman might not be reassured by the "safety
first" language in this section.